+++ src/Services/Sale/ShipmentPropertyValue/Result/UpdatedShipmentPropertyValueResult.php
@@ -1,0 +1,40 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Sale\ShipmentPropertyValue\Result;
+
+use Bitrix24\SDK\Core\Exceptions\BaseException;
+use Bitrix24\SDK\Core\Result\UpdatedItemResult;
+
+/**
+ * Updated result for sale.shipmentpropertyvalue.modify
+ */
+class UpdatedShipmentPropertyValueResult extends UpdatedItemResult
+{
+    /**
+     * @return ShipmentPropertyValueItemResult[]
+     * @throws BaseException
+     */
+    public function getPropertyValues(): array
+    {
+        $items = [];
+        foreach ($this->getCoreResponse()->getResponseData()->getResult()['propertyValues'] as $item) {
+            $items[] = new ShipmentPropertyValueItemResult($item);
+        }
+
+        return $items;
+    }
+
+    /**
+     * @throws BaseException
+     */
+    public function isSuccess(): bool
+    {
+        return isset($this->getCoreResponse()->getResponseData()->getResult()['propertyValues']);
+    }
+}